decorators

+++ HeadFirsts/DecoratorPattern/decorator/Milk.cpp
@@ -2,6 +2,6 @@
 
 Milk::Milk(Beverage *beverage) { this->beverage = beverage; }
 
-float Milk::cost() { return 0.20f + beverage->cost(); }
+double Milk::cost() { return 0.20f + beverage->cost(); }
 
 std::string Milk::getDescription() { return beverage->getDescription() + ", Milk"; }
